Move C interfaces
