+++ examples/call_qeq_from_c.c
@@ -3,8 +3,8 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include <oif/frontend_c/frontend_c.h>
-#include <oif/frontend_c/qeq.h>
+#include <oif/c_bindings.h>
+#include <oif/interfaces/qeq.h>
 
 
 char *parse_backend(int argc, char *argv[]) {
